src: favour whole int's for APIs
If the client needs to do math on the argument, and if the API
tries to use something smaller than an int (uint16_t or char),
then the arithmetic operators promote to ints, even if the
smaller values are unsigned, causing yet more warnings
and confusion.

So stick to int-sized arguments, whether unsigned or not, to avoid
this automatic promotion.