src: add request_id field to FastCGIRequest object, and update docs
It is possible, according to the protocol spec, that different
request_id packets can arrive interleaved between each other.
Therefore the application also needs to know which request_id
it is dealing with in each header, and how to handle the lifetime
of the object.

The id is now available in the request object, and the comments
updated.